feat(cuda): Fix 0 — host sampler wired into CUDA executor (greedy parity preserved)
Greedy path unchanged: params.is_greedy() still calls the existing device argmax helper, which still invokes ffi::argmax_cuda on the BF16 DeviceVec logits.

Temperature > 0 path: sample_cuda_token copies the BF16 logits DeviceVec to host f32 with DeviceVec::to_host(ctx), then calls infer_plan::sample_token(logits_host, params, position). Position uses the materialized logical length: prefill start_pos + tokens.len(), decode kv_seq_len + 1.

Helper LOC: 16. Repetition/frequency/presence penalties remain deferred because they need per-request generated-token history.

Verification: CUDARC_CUDA_VERSION=12060 cargo check -p infer-cuda --features cuda,no-cuda; CUDARC_CUDA_VERSION=12060 cargo clippy -p infer-cuda --features cuda,no-cuda -- -D warnings.

Note: the exact clippy gate also required removing one redundant f32 cast from infer-plan::sample_token; no behavior or seam change.

+++ crates/infer-plan/src/lib.rs
@@ -235,7 +235,7 @@ pub fn sample_token(logits: &[f32], params: &SamplingParams, position: u64) -> u
     let mut cand: Vec<(u32, f32)> = logits
         .iter()
         .enumerate()
-        .map(|(i, &l)| (i as u32, (((l - max) * inv_t) as f32).exp()))
+        .map(|(i, &l)| (i as u32, ((l - max) * inv_t).exp()))
         .collect();
     let sum: f32 = cand.iter().map(|(_, p)| *p).sum();
     if sum > 0.0 {
